hotel_api/requirements.txt
Django==5.2.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
reportlab==4.2.5
Pillow==11.1.0

